+++ trunk/src/codigo/TP taller/TP taller/Vista/Dibujables/RectanguloDibujable.h
@@ -21,7 +21,7 @@ public:
 	RectanguloDibujable();
 	~RectanguloDibujable();
 	void actualizar(Observable* observable);
-	void dibujar(SDL_Renderer* renderer, int corrimientoX,int corrimientoY, int escalaZoom, int anchoPx, int altoPx);
+	void dibujar(SDL_Renderer* renderer, int corrimientoX,int corrimientoY, float escalaZoom, int anchoPx, int altoPx);
 	virtual string serializar();
 	virtual void deserealizar(string aDeserealizar);
 };
